requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blockbuster==1.5.24
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
cryptography==43.0.3
cssselect2==0.8.0
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docker==7.1.0
dydantic==0.0.8
executing==2.2.0
fastjsonschema==2.21.1
fireworks-ai==0.15.12
fonttools==4.57.0
forbiddenfruit==0.1.4
fqdn==1.5.1
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
json5==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.29.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-community==0.3.21
langchain-core==0.3.51
langchain-fireworks==0.2.9
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.3.27
langgraph-api==0.0.43
langgraph-checkpoint==2.0.23
langgraph-checkpoint-sqlite==2.0.6
langgraph-cli==0.2.3
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langmem==0.0.19
langsmith==0.3.22
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.1.3
multidict==6.3.1
mypy-extensions==1.0.0
narwhals==1.34.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
# Editable Git install with no remote (react-agent==0.0.1)
-e /Users/timothyniko/Documents/lang/app
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
Send2Trash==1.8.3
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.1
streamlit==1.44.1
structlog==25.2.0
tavily-python==0.5.4
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trustcall==0.0.38
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
weasyprint==65.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wikipedia==1.4.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zopfli==0.2.3.post1
zstandard==0.23.0
